These modules are deleted: woo-admin-service woo-login-service wp-content-service woo-store-db
Better solution for collecting store data is implemented. Before woo-store-db was serving all store information, and now the system collects store endpoints using EurekaClient bean. Now, if the client is not registered it will no more search in that store. The search and match will only be done in active eureka clients with the application name "WOO-PRODUCT-SERVICE"